Removed SmartPhone endpoints, added and configured automapper for ShoesController, only for GETS, as an example
