+++ koszalka-api/DTO/BikeDTO.cs
@@ -1,0 +1,11 @@
+﻿namespace koszalka_api.Model
+{
+    public class BikeDTO : BaseEntity
+    {
+        public string? Name { get; set; }
+        public string? Description { get; set; }
+        public string? Price { get; set; }
+        public string? Model { get; set; }
+        public string? Brand { get; set; }
+    }
+}
