+++ koszalka-api/Interface/IShoesInterface.cs
@@ -6,8 +6,8 @@ namespace koszalka_api.Interface
 {
     public interface IShoesInterface
     {
-        Task<ActionResult<IEnumerable<Shoes>>> GetShoes();
-        Task<Shoes> GetShoe(long id);
+        Task<ActionResult<IEnumerable<ShoesDTO>>> GetShoes();
+        Task<ShoesDTO> GetShoe(long id);
         Task<Boolean> PutShoe(long id, Shoes shoes);
         Task<int> PostShoe(Shoes shoes);
         Task<int> DeleteShoe(long id);
