--- koszalka-api/Model/Smartphone.cs
@@ -1,9 +1,0 @@
-﻿namespace koszalka_api.Model
-{
-    public class Smartphone : BaseEntity
-    {
-        public string? Brand { get; set; }
-        public string? Price { get; set; }
-
-    }
-}
